修改评论接口和电影记录接口

+++ player-social/src/main/java/com/player/social/mapper/SocialMapper.java
@@ -23,7 +23,7 @@ public interface SocialMapper {
 
     Long saveLike(LikeEntity likeEntity);
 
-    LikeEntity getLikeById(int id);
+    LikeEntity getLikeById(Long id);
 
     Long deleteLike(Long relationId,String type,String userId);
